[LCR-104] refactor: 여러 client와의 분리를 위해 패키지 구조 변경
